requirements.txt
pytest
requests
pyyaml
sqlalchemy
psycopg2-binary
jsonschema
allure-pytest

